refactor/bitbucket test
