+++ .prettierignore
@@ -12,4 +12,5 @@ package.json
 test/_fixtures/
 test/datasource/**/_fixtures
 test/manager/**/_fixtures
+test/platform/**/_fixtures
 bin/yarn*
