+++ test/manager/gomod/update.spec.js
@@ -1,8 +1,8 @@
 const fs = require('fs');
 const goUpdate = require('../../../lib/manager/gomod/update');
 
-const gomod1 = fs.readFileSync('test/_fixtures/gomod/1/go.mod', 'utf8');
-const gomod2 = fs.readFileSync('test/_fixtures/gomod/2/go.mod', 'utf8');
+const gomod1 = fs.readFileSync('test/manager/gomod/_fixtures/1/go.mod', 'utf8');
+const gomod2 = fs.readFileSync('test/manager/gomod/_fixtures/2/go.mod', 'utf8');
 
 describe('manager/gomod/update', () => {
   describe('updateDependency', () => {
